requirements.txt
Flask
google-generativeai
python-dotenv
flask-cors
google.genai